+++ src/lib.rs
@@ -3,4 +3,5 @@ pub mod client;
 pub mod config;
 pub mod init;
 pub mod markdown;
+pub mod project;
 pub mod server;